refactored documentation in the swagger for the block of events, authentication and reviews
