+++ project/notifications/views.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any, Type
 
-from config.yasg import skip_param_query
+from config.openapi import skip_param_query
 from django.db.models.query import QuerySet
 from django.utils.decorators import (
     method_decorator,
